Created testing framework for hardware.
These tests will specifically be designed to test particular hardware
components with potential human intervention. The current design for
this framework is to have each test have its own firmware that will
execute on the desired board. The python scripts are there to compile
and upload each script as well as potentially passing variables to the
board if required. It will use a json configuration to manage switching
between the Arduino Nano and Arduino Mega.

Signed-off-by: Ethan Tola <ethanltola@gmail.com>