Refactored and completed new functionality.
It required a change in how the tests themselves are ran. The base dut
now contains the methods to compile, upload, and verify the test
configurations given to it.

To support multiple duts, a special suffix is used (similar to tests) to
match the _dut to the _tests suite. This way, its very easy to match the
board with the test it should execute. Multiple dut configurations will
run all the tests associated with them.

Signed-off-by: Ethan Tola <ethanltola@gmail.com>