Added screen switching data for Keypad Screen.
A working example of the screen switching was setup in the mega.ino.
There will be some weird stuff that will occur since the output data for
the Keypad screen isn't the same as the input.

Some work may have to be done to pass other metadata such as flags or
enums.

Signed-off-by: Ethan Tola <ethanltola@gmail.com>