Implemented baseline nano data flow.
Implemented a state machine to handle the LED components. Each function
still needs to be written out.

Putting everything in the main loop for now to ensure it all works. We
should probably breakout the current weight measurement to a timer so
the whole loop doesn't have to be delayed. Currently measuring at 20 Hz
(every 50 ms).

Signed-off-by: Ethan Tola <ethanltola@gmail.com>